Fix encoding of byte array attribute
This commit fixes the encoding of byte array attribute. According to the
code of p11-kit 0.24.x the size of the full data array is just the size
of the data, instead of the size of the length, 4, plus the size of the
data.

This version is compatible with p11-kit 0.24.x and 0.25.x..